requirements.txt
fastapi
uvicorn[standard]   # ASGI server runtime
pydantic>=2
pyyaml>=6           # YAML load/dump for specs & schemas